--- examples/README.md
@@ -1,51 +1,0 @@
-# Swap and Zap Examples
-
-This repository contains example scripts demonstrating the swapping and zapping functionality using the `symbiosis-js-sdk`.
-
-## Getting Started
-
-Follow the steps below to run these examples:
-
-### 1. Install Dependencies
-
-First, install all necessary dependencies using npm:
-
-```bash
-npm i
-```
-
-### 2. Run Examples
-
-To run an example, use the `npx tsx` command followed by the script name. Here are the available scripts:
-
-#### Swapping Native Tokens:
-
-```bash
-npx tsx swappingNative.ts
-```
-
-#### Swapping ERC20 Tokens:
-
-```bash
-npx tsx swappingErc20.ts
-```
-
-### Swapping TRC-20 (Tron) Tokens
-
-```bash
-npx tsx swappingTrc20.ts
-```
-
-#### Zapping Native Tokens to BTC:
-
-```bash
-npx tsx zappingNativeToBtc.ts
-```
-
-#### Zapping ERC20 Tokens to BTC:
-
-```bash
-npx tsx zappingErc20ToBtc.ts
-```
-
-> Choose the script that aligns with your use case and run it. Remember to configure the script appropriately to fit your environment, ensuring all necessary variables are correctly set.
