--- examples/package.json
@@ -1,23 +1,0 @@
-{
-    "name": "symbiosis-sdk-examples",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "type": "module",
-    "dependencies": {
-        "@ethersproject/constants": "^5.6.0",
-        "ethers": "^5.6.1",
-        "symbiosis-js-sdk": "latest"
-    },
-    "devDependencies": {
-        "@types/node": "^17.0.25",
-        "tsx": "^3.12.7",
-        "typescript": "^5.0.0"
-    }
-}
