--- examples/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "esnext",
-        "esModuleInterop": true,
-        "target": "esnext",
-        "moduleResolution": "node",
-        "outDir": "dist",
-        "rootDir": ".",
-        "typeRoots": ["./node_modules/@types"],
-        "baseUrl": ".",
-        "skipLibCheck": true,
-        "resolveJsonModule": true
-    },
-    "include": ["./**/*"],
-    "exclude": ["node_modules"],
-    "lib": ["es2015"]
-}
