+++ src/crosschain/chainUtils/evm.ts
@@ -12,7 +12,7 @@ import { BASES_TO_CHECK_TRADES_AGAINST, BIPS_BASE, CUSTOM_BASES } from '../const
 import { SdkError } from '../sdkError'
 import type { Symbiosis } from '../symbiosis'
 import { Field } from '../types'
-import { isChangellyNativeChainId } from '../swapExactIn/swapChangelly/constants'
+import { isChangellyNativeChainId } from '../swapExactIn/changellySwap/constants'
 import { isBtcChainId } from './btc'
 import { isQuaiChainId } from './quai'
 import { isSolanaChainId } from './solana'
