+++ src/crosschain/swapExactIn/chainFlipSwap/zappingCrossChainChainFlip.ts
@@ -5,7 +5,7 @@ import { TokenAmount } from '../../../entities'
 import { getMinAmount, isEvmChainId } from '../../chainUtils'
 import type { MulticallRouter } from '../../contracts'
 import { ChainFlipError } from '../../sdkError'
-import { BaseSwapping } from '../../swapping'
+import { BaseSwapping } from '../swapping'
 import type { OneInchProtocols } from '../../trade/oneInchTrade'
 import { TradeProvider } from '../../trade'
 import type { Address, EvmAddress, OmniPoolConfig, SwapExactInParams, SwapExactInResult } from '../../types'
