+++ src/crosschain/swapExactIn/index.ts
@@ -4,6 +4,6 @@ export {
     CHAIN_FLIP_FROM_SOLANA_TOKENS_OUT,
     CHAIN_FLIP_TO_SOLANA_TOKENS_IN,
     CHAIN_FLIP_TOKENS,
-} from './swapChainFlip'
+} from './chainFlipSwap'
 export * from './swapExactIn'
-export { THOR_TOKENS_IN } from './swapThorChain'
+export { THOR_TOKENS_IN } from './thorChainSwap'
