+++ src/crosschain/swapExactIn/toBtcSwap.ts
@@ -1,9 +1,9 @@
 import { isBtcChainId } from '../chainUtils'
 import { TradeProvider } from '../trade'
 import type { SwapExactInParams, SwapExactInResult } from '../types'
-import { btcChainFlipSwap } from './swapChainFlip'
-import { thorChainSwap } from './swapThorChain'
-import { burnSyntheticBtc } from './toBtc/burnSyntheticBtc'
+import { btcChainFlipSwap } from './chainFlipSwap'
+import { thorChainSwap } from './thorChainSwap'
+import { symbiosisBtcSwap } from './symbiosisBtcSwap/symbiosisBtcSwap'
 import { theBest } from './utils'
 
 export function isToBtcSwapSupported(context: SwapExactInParams): boolean {
@@ -15,7 +15,7 @@ export async function toBtcSwap(context: SwapExactInParams): Promise<SwapExactIn
 
     const promises: Promise<SwapExactInResult>[] = []
 
-    promises.push(burnSyntheticBtc(context))
+    promises.push(symbiosisBtcSwap(context))
 
     if (!disabledProviders?.includes(TradeProvider.THORCHAIN_BRIDGE)) {
         promises.push(thorChainSwap(context))
