--- src/crosschain/swapping/swappingMiddleware.ts
@@ -1,47 +1,0 @@
-import { withTracing } from '../tracing'
-import type { Address, SwapExactInParams, SwapExactInResult } from '../types'
-import { BaseSwapping } from './baseSwapping'
-import { Swapping } from './swapping'
-
-export class SwappingMiddleware extends BaseSwapping {
-    protected middlewareAddress!: Address
-    protected middlewareData!: string
-    protected middlewareOffset!: number
-
-    @withTracing({
-        onCall: function (_) {
-            return {
-                'omniPool.chainId': this.omniPoolConfig.chainId,
-                'omniPool.address': this.omniPoolConfig.address,
-            }
-        },
-    })
-    public async exactIn(params: Omit<SwapExactInParams, 'symbiosis'>): Promise<SwapExactInResult> {
-        const { middlewareCall } = params
-        if (!middlewareCall) {
-            const { symbiosis, omniPoolConfig } = this
-            const swapping = new Swapping(symbiosis, omniPoolConfig)
-            return swapping.doExactIn(params)
-        }
-
-        const { address, data, offset } = middlewareCall
-
-        this.middlewareAddress = address
-        this.middlewareData = data
-        this.middlewareOffset = offset
-
-        return this.doExactIn(params)
-    }
-
-    protected finalReceiveSide(): Address {
-        return this.middlewareAddress
-    }
-
-    protected finalCalldata(): string | [] {
-        return this.middlewareData
-    }
-
-    protected finalOffset(): number {
-        return this.middlewareOffset
-    }
-}
