+++ test/crosschain/swapExactIn/crosschainSwap/bestPoolSwapping.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from 'vitest'
 
 import { ChainId, GAS_TOKEN, Symbiosis, Token, WETH } from '../../../../src'
-import type { Route } from '../../../../src/crosschain/swapExactIn/crosschainSwap/bestPoolSwapping'
-import { getRoutes } from '../../../../src/crosschain/swapExactIn/crosschainSwap/bestPoolSwapping'
+import type { Route } from '../../../../src/crosschain/swapExactIn/crossChainSwap/bestPoolSwapping'
+import { getRoutes } from '../../../../src/crosschain/swapExactIn/crossChainSwap/bestPoolSwapping'
 
 const symbiosis = new Symbiosis('mainnet', 'test')
